Fix Bluetooth UI freezes
