+++ cmake/PrintSummary.cmake
@@ -95,7 +95,6 @@ OptionOutput("Unity build                           " QGC_UNITY_BUILD)
 OptionOutput("Enable testing                        " QGC_BUILD_TESTING)
 OptionOutput("Enable QML debugging                  " QGC_DEBUG_QML)
 OptionOutput("Enable QML linting                    " QGC_ENABLE_QMLLINT)
-OptionOutput("Enable Bluetooth links                " QGC_ENABLE_BLUETOOTH)
 OptionOutput("Enable ZeroConf compatibility         " QGC_ZEROCONF_ENABLED)
 OptionOutput("Disable serial links                  " QGC_NO_SERIAL_LINK)
 OptionOutput("Enable UVC devices                    " QGC_ENABLE_UVC)
