+++ src/Comms/Bluetooth/BluetoothConfiguration.h
@@ -1,30 +1,19 @@
 #pragma once
 
 #include "LinkConfiguration.h"
-#include "LinkInterface.h"
 
 #include <QtBluetooth/QBluetoothAddress>
 #include <QtBluetooth/QBluetoothDeviceDiscoveryAgent>
 #include <QtBluetooth/QBluetoothDeviceInfo>
-#include <QtBluetooth/QBluetoothSocket>
+#include <QtBluetooth/QBluetoothHostInfo>
 #include <QtBluetooth/QBluetoothLocalDevice>
-#include <QtBluetooth/QBluetoothServiceInfo>
-#include <QtBluetooth/QBluetoothServiceDiscoveryAgent>
-#include <QtBluetooth/QLowEnergyController>
-#include <QtBluetooth/QLowEnergyService>
-#include <QtBluetooth/QLowEnergyCharacteristic>
+#include <QtCore/QFutureWatcher>
 #include <QtCore/QList>
 #include <QtCore/QLoggingCategory>
 #include <QtCore/QPointer>
-#include <QtCore/QQueue>
 #include <QtCore/QString>
-#include <QtCore/QTimer>
 #include <QtCore/QVariantList>
 #include <QtQmlIntegration/QtQmlIntegration>
-
-#include <atomic>
-
-class QThread;
 
 Q_DECLARE_LOGGING_CATEGORY(BluetoothLinkLog)
 Q_DECLARE_LOGGING_CATEGORY(BluetoothLinkVerboseLog)
@@ -39,7 +28,6 @@ class BluetoothConfiguration : public LinkConfiguration
     Q_PROPERTY(BluetoothMode mode            READ mode            WRITE setMode           NOTIFY modeChanged)
     Q_PROPERTY(QString       deviceName      READ deviceName                              NOTIFY deviceChanged)
     Q_PROPERTY(QString       address         READ address                                 NOTIFY deviceChanged)
-    Q_PROPERTY(QStringList   nameList        READ nameList                                NOTIFY nameListChanged)
     Q_PROPERTY(QVariantList  devicesModel    READ devicesModel                            NOTIFY devicesModelChanged)
     Q_PROPERTY(bool          scanning        READ scanning                                NOTIFY scanningChanged)
     Q_PROPERTY(QString       serviceUuid     READ serviceUuid     WRITE setServiceUuid    NOTIFY serviceUuidChanged)
@@ -49,6 +37,7 @@ class BluetoothConfiguration : public LinkConfiguration
     Q_PROPERTY(qint16        selectedRssi    READ selectedRssi                            NOTIFY selectedRssiChanged)
     Q_PROPERTY(bool          adapterAvailable READ isAdapterAvailable                     NOTIFY adapterStateChanged)
     Q_PROPERTY(bool          adapterPoweredOn READ isAdapterPoweredOn                     NOTIFY adapterStateChanged)
+    Q_PROPERTY(bool          adapterDiscoverable READ isDiscoverable                      NOTIFY adapterStateChanged)
     Q_PROPERTY(QString       adapterName     READ getAdapterName                          NOTIFY adapterStateChanged)
     Q_PROPERTY(QString       adapterAddress  READ getAdapterAddress                       NOTIFY adapterStateChanged)
     Q_PROPERTY(QString       hostMode        READ getHostMode                             NOTIFY adapterStateChanged)
@@ -73,18 +62,20 @@ public:
     Q_INVOKABLE void requestPairing(const QString &address);
     Q_INVOKABLE void removePairing(const QString &address);
     Q_INVOKABLE QString getPairingStatus(const QString &address) const;
+    [[nodiscard]] Q_INVOKABLE bool isPaired(const QString &address) const;
+    [[nodiscard]] Q_INVOKABLE bool isDiscoverable() const;
 
     // Adapter information
-    Q_INVOKABLE bool isAdapterAvailable() const;
+    [[nodiscard]] Q_INVOKABLE bool isAdapterAvailable() const;
     Q_INVOKABLE QString getAdapterAddress() const;
     Q_INVOKABLE QString getAdapterName() const;
-    Q_INVOKABLE bool isAdapterPoweredOn() const;
+    [[nodiscard]] Q_INVOKABLE bool isAdapterPoweredOn() const;
     Q_INVOKABLE QVariantList getAllPairedDevices() const;
     Q_INVOKABLE QVariantList getConnectedDevices() const;
     Q_INVOKABLE void powerOnAdapter();
     Q_INVOKABLE void powerOffAdapter();
     Q_INVOKABLE void setAdapterDiscoverable(bool discoverable);
-    Q_INVOKABLE QVariantList getAllAvailableAdapters() const;
+    Q_INVOKABLE QVariantList getAllAvailableAdapters();
     Q_INVOKABLE void selectAdapter(const QString &address);
     Q_INVOKABLE QString getHostMode() const;
 
@@ -101,7 +92,6 @@ public:
     const QBluetoothDeviceInfo& device() const { return _device; }
     QString deviceName() const { return _device.name(); }
     QString address() const { return _device.address().toString(); }
-    const QStringList& nameList() const { return _nameList; }
     QVariantList devicesModel() const;
     bool scanning() const;
     qint16 connectedRssi() const { return _connectedRssi; }
@@ -119,7 +109,7 @@ public:
     const QBluetoothUuid& readCharacteristicUuid() const { return _readCharacteristicUuid; }
     const QBluetoothUuid& writeCharacteristicUuid() const { return _writeCharacteristicUuid; }
 
-    static bool isBluetoothAvailable();
+    [[nodiscard]] static bool isBluetoothAvailable();
 
     // Known BLE UART-like service UUIDs
     static inline const QBluetoothUuid NORDIC_UART_SERVICE{QStringLiteral("6e400001-b5a3-f393-e0a9-e50e24dcca9e")};
@@ -131,7 +121,6 @@ public:
 signals:
     void modeChanged();
     void deviceChanged();
-    void nameListChanged();
     void devicesModelChanged();
     void scanningChanged();
     void serviceUuidChanged();
@@ -158,6 +147,10 @@ private slots:
 
 private:
     void _initDeviceDiscoveryAgent();
+    void _refreshAvailableAdaptersAsync();
+    void _updateAvailableAdapters(const QList<QBluetoothHostInfo> &hosts);
+    bool _ensureLocalDevice();
+    void _requestHostMode(QBluetoothLocalDevice::HostMode mode);
     void _updateDeviceList();
     void _applySelectedDevice(const QBluetoothDeviceInfo &info);
     bool _createLocalDevice(const QBluetoothAddress &address);
@@ -167,161 +160,21 @@ private:
     BluetoothMode _mode = BluetoothMode::ModeClassic;
     QBluetoothDeviceInfo _device;
     QList<QBluetoothDeviceInfo> _deviceList;
-    QStringList _nameList;
+    QVariantList _devicesModelCache;
     QPointer<QBluetoothDeviceDiscoveryAgent> _deviceDiscoveryAgent;
     QPointer<QBluetoothLocalDevice> _localDevice;
+    QVariantList _availableAdapters;
+    QPointer<QFutureWatcher<QList<QBluetoothHostInfo>>> _adapterWatcher;
+    bool _adapterEnumerationInProgress = false;
+    bool _hostModeRequestPending = false;
+    bool _deferredPowerOnFixPending = false;
+    QBluetoothLocalDevice::HostMode _requestedHostMode = QBluetoothLocalDevice::HostPoweredOff;
+    bool _hasQueuedHostModeRequest = false;
+    QBluetoothLocalDevice::HostMode _queuedHostMode = QBluetoothLocalDevice::HostPoweredOff;
     qint16 _connectedRssi = 0;
 
     // BLE specific - Default to Nordic UART Service
     QBluetoothUuid _serviceUuid{NORDIC_UART_SERVICE};
     QBluetoothUuid _readCharacteristicUuid{NORDIC_UART_RX_CHAR};
     QBluetoothUuid _writeCharacteristicUuid{NORDIC_UART_TX_CHAR};
-};
-
-/*===========================================================================*/
-
-class BluetoothWorker : public QObject
-{
-    Q_OBJECT
-
-public:
-    explicit BluetoothWorker(const BluetoothConfiguration *config, QObject *parent = nullptr);
-    ~BluetoothWorker() override;
-
-    bool isConnected() const;
-
-signals:
-    void connected();
-    void disconnected();
-    void errorOccurred(const QString &errorString);
-    void dataReceived(const QByteArray &data);
-    void dataSent(const QByteArray &data);
-    void rssiUpdated(qint16 rssi);
-
-public slots:
-    void setupConnection();
-    void connectLink();
-    void disconnectLink();
-    void writeData(const QByteArray &data);
-
-private slots:
-    // Classic Bluetooth slots
-    void _onSocketConnected();
-    void _onSocketDisconnected();
-    void _onSocketReadyRead();
-    void _onSocketBytesWritten(qint64 bytes);
-    void _onSocketErrorOccurred(QBluetoothSocket::SocketError socketError);
-    void _onClassicServiceDiscovered(const QBluetoothServiceInfo &serviceInfo);
-    void _onClassicServiceDiscoveryFinished();
-    void _onClassicServiceDiscoveryCanceled();
-    void _onClassicServiceDiscoveryError(QBluetoothServiceDiscoveryAgent::Error error);
-
-    // BLE slots
-    void _onControllerConnected();
-    void _onControllerDisconnected();
-    void _onControllerErrorOccurred(QLowEnergyController::Error error);
-    void _onServiceDiscovered(const QBluetoothUuid &uuid);
-    void _onServiceDiscoveryFinished();
-    void _onServiceStateChanged(QLowEnergyService::ServiceState state);
-    void _onCharacteristicChanged(const QLowEnergyCharacteristic &characteristic, const QByteArray &value);
-    void _onCharacteristicRead(const QLowEnergyCharacteristic &characteristic, const QByteArray &value);
-    void _onCharacteristicWritten(const QLowEnergyCharacteristic &characteristic, const QByteArray &value);
-    void _onDescriptorRead(const QLowEnergyDescriptor &descriptor, const QByteArray &value);
-    void _onDescriptorWritten(const QLowEnergyDescriptor &descriptor, const QByteArray &value);
-    void _onServiceError(QLowEnergyService::ServiceError error);
-
-    // Common slots
-    void _reconnectTimeout();
-    void _serviceDiscoveryTimeout();
-
-private:
-    void _setupClassicSocket();
-    void _startClassicServiceDiscovery();
-    void _setupBleController();
-    void _recreateBleController();
-    void _setupBleService();
-    void _discoverServiceDetails();
-    void _enableNotifications();
-    void _writeBleData(const QByteArray &data);
-    void _writeClassicData(const QByteArray &data);
-    void _findCharacteristics();
-    void _processNextBleWrite();
-    void _clearBleWriteQueue();
-
-    // Cached config values (copied at construction for thread safety)
-    const BluetoothConfiguration::BluetoothMode _mode;
-    const QBluetoothDeviceInfo _device;
-    const QString _serviceUuid;
-    const QBluetoothUuid _readCharacteristicUuid;
-    const QBluetoothUuid _writeCharacteristicUuid;
-
-    // Classic Bluetooth
-    QPointer<QBluetoothSocket> _socket;
-
-    // BLE
-    QPointer<QLowEnergyController> _controller;
-    QPointer<QLowEnergyService> _service;
-    QLowEnergyCharacteristic _readCharacteristic;
-    QLowEnergyCharacteristic _writeCharacteristic;
-    int _mtu = DEFAULT_ATT_MTU;
-    qint16 _rssi = 0;
-    QPointer<QTimer> _rssiTimer;
-    QQueue<QByteArray> _bleWriteQueue;
-    QByteArray _currentBleWrite;
-    bool _bleWriteInProgress = false;
-
-    // Common
-    QPointer<QTimer> _reconnectTimer;
-    QPointer<QTimer> _serviceDiscoveryTimer;
-    std::atomic<bool> _intentionalDisconnect{false};
-    std::atomic<bool> _connected{false};
-    std::atomic<int> _reconnectAttempts{0};
-    int _consecutiveFailures = 0;
-    static constexpr int MAX_RECONNECT_ATTEMPTS = 10;
-    static constexpr int MAX_BLE_QUEUE_SIZE = 100;
-    static constexpr int RSSI_POLL_INTERVAL_MS = 10000;
-    static constexpr int RECONNECT_BASE_INTERVAL_MS = 5000;
-    static constexpr int MAX_RECONNECT_INTERVAL_MS = 60000;
-    static constexpr int SERVICE_DISCOVERY_TIMEOUT_MS = 30000;
-    static constexpr int MAX_CONSECUTIVE_FAILURES = 3;
-    QPointer<QBluetoothServiceDiscoveryAgent> _classicDiscovery;
-    QBluetoothServiceInfo _classicDiscoveredService;
-
-    // BLE packet size constraints
-    static constexpr int BLE_MIN_PACKET_SIZE = 20;
-    static constexpr int BLE_MAX_PACKET_SIZE = 512;
-    static constexpr int DEFAULT_ATT_MTU = 23;
-};
-
-/*===========================================================================*/
-
-class BluetoothLink : public LinkInterface
-{
-    Q_OBJECT
-
-public:
-    explicit BluetoothLink(SharedLinkConfigurationPtr &config, QObject *parent = nullptr);
-    ~BluetoothLink() override;
-
-    bool isConnected() const override;
-    void disconnect() override;
-
-private slots:
-    void _writeBytes(const QByteArray &bytes) override;
-    void _onConnected();
-    void _onDisconnected();
-    void _onErrorOccurred(const QString &errorString);
-    void _onDataReceived(const QByteArray &data);
-    void _onDataSent(const QByteArray &data);
-    void _onRssiUpdated(qint16 rssi);
-
-private:
-    bool _connect() override;
-    void _checkPermission();
-    void _handlePermissionStatus(Qt::PermissionStatus permissionStatus);
-
-    BluetoothConfiguration *_bluetoothConfig = nullptr;
-    QPointer<BluetoothWorker> _worker;
-    QPointer<QThread> _workerThread;
-    std::atomic<bool> _disconnectedEmitted{false};
 };
