+++ src/Comms/LinkConfiguration.h
@@ -76,9 +76,7 @@ public:
 #endif
         TypeUdp,        ///< UDP Link
         TypeTcp,        ///< TCP Link
-#ifdef QGC_ENABLE_BLUETOOTH
         TypeBluetooth,  ///< Bluetooth Link
-#endif
 #ifdef QT_DEBUG
         TypeMock,       ///< Mock Link for Unitesting
 #endif
