+++ test/Comms/Bluetooth/BluetoothEmulatedAdapterTest.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "UnitTest.h"
+
+class BluetoothEmulatedAdapterTest : public UnitTest
+{
+    Q_OBJECT
+
+private slots:
+    void _testOptionalEmulatorHarness();
+};
